requirement.txt
asyncio==3.4.3
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
idna==3.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
paramiko==2.9.1
protobuf==3.19.1
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
